pyproject.toml
[project]
name = "auto-review"
version = "0.1.0"
description = "Automated code review workflow: Gemini reviews MR diffs, Claude Code fixes issues"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "flask>=3.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
auto-review = "src.runner:main"
